tutto rotto

+++ lib/libft/ft_strdup.c
@@ -6,12 +6,12 @@
 /*   By: anovelli <anovelli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/12 19:32:15 by anovelli          #+#    #+#             */
-/*   Updated: 2022/11/17 10:25:11 by anovelli         ###   ########.fr       */
+/*   Updated: 2022/12/05 17:49:06 by anovelli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"libft.h"
-
+# include <stdio.h>
 char	*ft_strdup(const char *s1)
 {
 	int		i;
@@ -21,7 +21,7 @@ char	*ft_strdup(const char *s1)
 	i = 0;
 	if (!s1)
 		return (NULL);
-	len = ft_strlen(s1);
+	len = (int)ft_strlen(s1);
 	dest = malloc(sizeof(char) * (len + 1));
 	while (s1[i] != '\0')
 	{
